==> 01_TCGA_miRNA_Processing.R <==
SCRIPT 01: TCGA Pan-Cancer miRNA Data ProcessingPURPOSE: Download, clean, and normalize all 33 TCGA miRNA datasets.RUNTIME: ~30-45 minutesOUTPUT: vsd_miRNA_Normalized.rds--- 1. LOAD LIBRARIES ---(Install if missing)BiocManager::install("TCGAbiolinks")BiocManager::install("DESeq2")library(TCGAbiolinks)library(DESeq2)--- 2. QUERY TCGA PROJECTS ---print("Querying all TCGA projects...")projects <- getGDCprojects()tcga_projects <- projects[startsWith(projects$project_id, "TCGA-"), ]--- 3. QUERY miRNA DATA (ALL 33 CANCERS) ---print("Querying miRNA data for all 33 cancers...")query_mirna <- GDCquery(project = tcga_projects$project_id,data.category = "Transcriptome Profiling",data.type = "miRNA Expression Quantification",workflow.type = "BCGSC miRNA Profiling")--- 4. DOWNLOAD DATA ---(This step downloads all files, ~550MB. May take time)print("Downloading all miRNA data...")GDCdownload(query = query_mirna, method = "api", files.per.chunk = 100)--- 5. PREPARE DATA (Long Step: ~15-20 min) ---(Reads all 11,000+ files into one object)print("Preparing (GDCprepare) all miRNA data. This will take ~15-20 min...")data_prepared <- GDCprepare(query = query_mirna)--- 6. BUILD CLEAN COUNT MATRIX ---print("Building clean count matrix...")row_names <- data_prepared$miRNA_IDcount_columns <- grepl("^read_count", colnames(data_prepared))count_matrix <- data_prepared[, count_columns]rownames(count_matrix) <- row_namescolnames(count_matrix) <- sub("read_count_", "", colnames(count_matrix))--- 7. BUILD CLEAN METADATA ---print("Building clean metadata...")results_table <- getResults(query_mirna)metadata <- data.frame(barcode_long = colnames(count_matrix),project_id = results_table$project,sample_type = results_table$sample_type)rownames(metadata) <- metadata$barcode_long--- 8. FILTER NON-PRIMARY SAMPLES ---print("Filtering non-primary/normal samples...")types_to_keep <- c("Primary Tumor", "Solid Tissue Normal", "Primary Blood Derived Cancer - Peripheral Blood")clean_metadata <- metadata[metadata$sample_type %in% types_to_keep, ]clean_count_matrix <- count_matrix[, rownames(clean_metadata)]--- 9. FILTER LOW-COUNT MIRNAS ---print("Filtering low-count miRNAs...")miRNA_sums <- rowSums(clean_count_matrix)rows_to_keep <- miRNA_sums > 1000 # Keep miRNAs with >1000 total countsfinal_matrix <- clean_count_matrix[rows_to_keep, ]--- 10. CREATE DESEQ OBJECT ---print("Creating DESeq object...")dds <- DESeqDataSetFromMatrix(countData = final_matrix,colData = clean_metadata,design = ~ 1)--- 11. NORMALIZE DATA (Long Step: ~10-15 min) ---print("Running VST normalization... This will take ~10-15 min...")(We use the direct function to avoid 'nsub' errors)vsd <- varianceStabilizingTransformation(dds, blind = TRUE)--- 12. SAVE FINAL NORMALIZED OBJECT ---print("Saving normalized data object (vsd)...")saveRDS(vsd, "vsd_miRNA_Normalized.rds")print("--- SCRIPT 01 COMPLETE ---")
